Fix the index bug due to the use of java faker
- Closes #77

+++ evaluations-service-core/evaluations-service-models/src/test/java/ar/edu/itba/cep/evaluations_service/models/ExerciseSolutionResultTest.java
@@ -145,7 +145,7 @@ class ExerciseSolutionResultTest {
      */
     private static ExerciseSolutionResult.Result validResult() {
         final var results = ExerciseSolutionResult.Result.values();
-        final var randomIndex = Faker.instance().number().numberBetween(0, results.length);
+        final var randomIndex = (int) Faker.instance().number().numberBetween(0L, results.length);
         return results[randomIndex];
     }
 }
